Add if checks to user data to avoid re-running initialization after restart

+++ infrastructure/aws-ec2-gpu/ec2-instance.tf
@@ -49,44 +49,54 @@ resource "aws_instance" "ec2_instance" {
   user_data = <<-EOF
     #!/bin/bash
     
-    sudo NEEDRESTART_MODE=a apt-get update -y
-    # Git and build essentials (autorestart daemons if needed)
-    sudo NEEDRESTART_MODE=a apt-get install -y git build-essential
-    touch /home/ubuntu/_build_essential
+    if [[ ! -f /home/ubuntu/_build_essential ]]; then
+      sudo NEEDRESTART_MODE=a apt-get update -y
+      # Git and build essentials (autorestart daemons if needed)
+      sudo NEEDRESTART_MODE=a apt-get install -y git build-essential
+      touch /home/ubuntu/_build_essential
+    fi
 
     # Docker
-    sudo NEEDRESTART_MODE=a apt install -y apt-transport-https ca-certificates curl gnupg-agent software-properties-common
-    curl -fsSL https://download.docker.com/linux/ubuntu/gpg | sudo NEEDRESTART_MODE=a apt-key add -
-    sudo NEEDRESTART_MODE=a add-apt-repository -y "deb [arch=amd64] https://download.docker.com/linux/ubuntu $(lsb_release -cs) stable"
-    sudo NEEDRESTART_MODE=a apt update
-    sudo NEEDRESTART_MODE=a apt install -y docker-ce docker-ce-cli containerd.io
-    # Enable for next restart(s)
-    sudo NEEDRESTART_MODE=a systemctl enable docker
-    sudo NEEDRESTART_MODE=a usermod -a -G docker ubuntu
-    touch /home/ubuntu/_docker
+    if [[ ! -f /home/ubuntu/_docker ]]; then
+      sudo NEEDRESTART_MODE=a apt install -y apt-transport-https ca-certificates curl gnupg-agent software-properties-common
+      curl -fsSL https://download.docker.com/linux/ubuntu/gpg | sudo NEEDRESTART_MODE=a apt-key add -
+      sudo NEEDRESTART_MODE=a add-apt-repository -y "deb [arch=amd64] https://download.docker.com/linux/ubuntu $(lsb_release -cs) stable"
+      sudo NEEDRESTART_MODE=a apt update
+      sudo NEEDRESTART_MODE=a apt install -y docker-ce docker-ce-cli containerd.io
+      # Enable for next restart(s)
+      sudo NEEDRESTART_MODE=a systemctl enable docker
+      sudo NEEDRESTART_MODE=a usermod -a -G docker ubuntu
+      touch /home/ubuntu/_docker
+    fi
 
     # Docker-compose
-    sudo NEEDRESTART_MODE=a curl -L https://github.com/docker/compose/releases/download/v2.24.0/docker-compose-$(uname -s)-$(uname -m) -o /usr/local/bin/docker-compose
-    sudo NEEDRESTART_MODE=a chmod +x /usr/local/bin/docker-compose
-    touch /home/ubuntu/_docker_compose
+    if [[ ! -f /home/ubuntu/_docker_compose ]]; then
+      sudo NEEDRESTART_MODE=a curl -L https://github.com/docker/compose/releases/download/v2.24.0/docker-compose-$(uname -s)-$(uname -m) -o /usr/local/bin/docker-compose
+      sudo NEEDRESTART_MODE=a chmod +x /usr/local/bin/docker-compose
+      touch /home/ubuntu/_docker_compose
+    fi
 
     # Install Nvidia drivers
-    BASE_URL=https://us.download.nvidia.com/tesla
-    DRIVER_VERSION=535.129.03
-    curl -fSsl -O $BASE_URL/$DRIVER_VERSION/NVIDIA-Linux-x86_64-$DRIVER_VERSION.run
-    chmod +x NVIDIA-Linux-x86_64-$DRIVER_VERSION.run 
-    sudo NEEDRESTART_MODE=a sh NVIDIA-Linux-x86_64-$DRIVER_VERSION.run -qs
-    touch /home/ubuntu/_cuda
+    if [[ ! -f /home/ubuntu/_cuda ]]; then
+      BASE_URL=https://us.download.nvidia.com/tesla
+      DRIVER_VERSION=535.129.03
+      curl -fSsl -O $BASE_URL/$DRIVER_VERSION/NVIDIA-Linux-x86_64-$DRIVER_VERSION.run
+      chmod +x NVIDIA-Linux-x86_64-$DRIVER_VERSION.run 
+      sudo NEEDRESTART_MODE=a sh NVIDIA-Linux-x86_64-$DRIVER_VERSION.run -qs
+      touch /home/ubuntu/_cuda
+    fi
 
     # Nvidia container toolkit
-    curl -fsSL https://nvidia.github.io/libnvidia-container/gpgkey | sudo NEEDRESTART_MODE=a gpg --dearmor -o /usr/share/keyrings/nvidia-container-toolkit-keyring.gpg
-    curl -s -L https://nvidia.github.io/libnvidia-container/stable/deb/nvidia-container-toolkit.list | sed 's#deb https://#deb [signed-by=/usr/share/keyrings/nvidia-container-toolkit-keyring.gpg] https://#g' | sudo NEEDRESTART_MODE=a tee /etc/apt/sources.list.d/nvidia-container-toolkit.list
-    sudo NEEDRESTART_MODE=a apt-get update
-    sudo NEEDRESTART_MODE=a apt-get install -y nvidia-container-toolkit
-    sudo NEEDRESTART_MODE=a systemctl restart docker
-    sudo nvidia-ctk runtime configure --runtime=docker
-    sudo systemctl restart docker
-    touch /home/ubuntu/_toolkit
+    if [[ ! -f /home/ubuntu/_toolkit ]]; then
+      curl -fsSL https://nvidia.github.io/libnvidia-container/gpgkey | sudo NEEDRESTART_MODE=a gpg --dearmor -o /usr/share/keyrings/nvidia-container-toolkit-keyring.gpg
+      curl -s -L https://nvidia.github.io/libnvidia-container/stable/deb/nvidia-container-toolkit.list | sed 's#deb https://#deb [signed-by=/usr/share/keyrings/nvidia-container-toolkit-keyring.gpg] https://#g' | sudo NEEDRESTART_MODE=a tee /etc/apt/sources.list.d/nvidia-container-toolkit.list
+      sudo NEEDRESTART_MODE=a apt-get update
+      sudo NEEDRESTART_MODE=a apt-get install -y nvidia-container-toolkit
+      sudo NEEDRESTART_MODE=a systemctl restart docker
+      sudo nvidia-ctk runtime configure --runtime=docker
+      sudo systemctl restart docker
+      touch /home/ubuntu/_toolkit
+    fi
 
     reboot
   EOF
